Move MonitoredViews to core:ui
