+++ core/ui/src/main/java/com/buzbuz/smartautoclicker/core/ui/monitoring/ViewMonitor.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.buzbuz.smartautoclicker.feature.tutorial.data.monitoring
+package com.buzbuz.smartautoclicker.core.ui.monitoring
 
 import android.graphics.Point
 import android.graphics.Rect
