--- feature/tutorial/src/main/java/com/buzbuz/smartautoclicker/feature/tutorial/domain/TutorialMonitoredViewsManager.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright (C) 2023 Kevin Buzeau
- * 
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.buzbuz.smartautoclicker.feature.tutorial.domain
-
-import android.view.View
-import com.buzbuz.smartautoclicker.feature.tutorial.data.monitoring.MonitoredViewsManager
-import com.buzbuz.smartautoclicker.feature.tutorial.data.monitoring.ViewPositioningType
-import com.buzbuz.smartautoclicker.feature.tutorial.domain.model.monitoring.TutorialMonitoredViewType
-
-interface TutorialMonitoredViewsManager {
-
-    companion object {
-        fun getInstance(): TutorialMonitoredViewsManager = MonitoredViewsManager.getInstance()
-    }
-
-    fun attach(
-        type: TutorialMonitoredViewType,
-        monitoredView: View,
-        positioningType: ViewPositioningType = ViewPositioningType.WINDOW,
-    )
-
-    fun detach(type: TutorialMonitoredViewType)
-
-    fun notifyClick(type: TutorialMonitoredViewType)
-}
